cleanup and run tests
